feat: Add MIT License
This commit introduces the LICENSE.md file with the MIT License.

The license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, subject to the inclusion of the copyright notice and license in all copies or substantial portions of the software.

The software is provided "AS IS", without warranty of any kind.